Move more into PlatformIO
