+++ core/src/main/scala/org/bykn/bosatsu/tool/ExitCode.scala
@@ -1,0 +1,7 @@
+package org.bykn.bosatsu.tool
+
+sealed abstract class ExitCode(val toInt: Int)
+object ExitCode {
+  case object Success extends ExitCode(0)
+  case object Error extends ExitCode(1)
+}
